test: move it to integration tests
