+++ tokio-console/tests/ui/cli-ui.stdout
@@ -1,6 +1,6 @@
 The Tokio console: a debugger for async Rust.
 
-Usage: tokio-console [OPTIONS] [TARGET_ADDR] [COMMAND]
+Usage: tokio-console[EXE] [OPTIONS] [TARGET_ADDR] [COMMAND]
 
 Commands:
   gen-config
